requirements.txt
opencv-python-headless
Pillow
python-telegram-bot
numpy
streamlink
python-dotenv
sentry-sdk